Refactor the HookSystem
This Commit uses the -fpatchable-function-entry=N[,M] parameter of GCC and Clang
to inject 12 bytes of NOPs at the start of each function. The HookSystem the
just injects the JMP to the hook into this section at the start of the function.

Unhooking is done by replacing the first 12 bytes of the function with NOPs again.